+++ src/ft_mem/ft_calloc.c
@@ -6,16 +6,16 @@
 /*   By: evdalmas <evdalmas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/08 11:21:17 by evdalmas          #+#    #+#             */
-/*   Updated: 2024/11/08 11:25:40 by evdalmas         ###   ########.fr       */
+/*   Updated: 2025/01/08 17:07:20 by evdalmas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../../includes/libft.h"
 
-void	*ft_calloc(int elementCount, int elementSize)
+void	*ft_calloc(size_t elementCount, size_t elementSize)
 {
 	void	*ptr;
-	int		i;
+	size_t		i;
 
 	ptr = malloc(elementCount * elementSize);
 	if (ptr == NULL)
